get rid of old routing framework
