+++ web/package.json
@@ -25,7 +25,6 @@
     "@types/d3": "^5.0.0",
     "@types/d3-shape": "^1.2.2",
     "@types/jquery": "^3.3.0",
-    "@types/react-router-dom": "^4.3.0",
     "@types/underscore": "^1.8.7",
     "ajv": "^6.2.0",
     "axios": "^0.18.0",
@@ -41,10 +40,7 @@
     "jquery": "^3.3.1",
     "loglevel": "^1.6.1",
     "react-codemirror2": "^5.1.0",
-    "react-router-dom": "^4.3.1",
     "reflect-metadata": "^0.1.12",
-    "restyped-express-async": "^0.6.1",
-    "routing-controllers": "^0.7.6",
     "sqlite3": "^4.0.2",
     "topside": "^0.6.4",
     "tslib": "^1.9.3",
@@ -67,9 +63,9 @@
     "@types/uuid": "^3.4.3",
     "@types/webpack": "^4.4.8",
     "@types/webpack-env": "^1.13.5",
+    "class-transformer": "^0.1.9",
     "cors": "^2.8.4",
     "jest": "^23.4.2",
-    "restyped-axios": "^1.3.2",
     "ts-jest": "^23.0.1",
     "ts-loader": "^4.4.2",
     "ts-node": "^7.0.0",
